Initial datatype refactor
