+++ lumps/datatypes/brush/brush.go
@@ -1,0 +1,7 @@
+package brush
+
+type Brush struct {
+	FirstSize int32
+	NumSides  int32
+	Contents int32
+}
