+++ lumps/datatypes/brushside/brushside.go
@@ -1,0 +1,9 @@
+package brushside
+
+
+type BrushSide struct {
+	PlaneNum uint16
+	TexInfo int16
+	DispInfo int16
+	Bevel int16
+}
