+++ lumps/datatypes/game/game.go
@@ -1,17 +1,17 @@
-package lumpdata
+package game
 
 
-type GameHeader struct {
+type Header struct {
 	LumpCount int32
-	GameLumps []GameLump // Slice length must equal lumpCount. Validation to be added
+	GameLumps []LumpDef // Slice length must equal lumpCount. Validation to be added
 }
-func (header GameHeader) SetLumpCount(num int32) GameHeader{
+func (header Header) SetLumpCount(num int32) Header{
 	header.LumpCount = num
-	header.GameLumps = make([]GameLump, header.LumpCount)
+	header.GameLumps = make([]LumpDef, header.LumpCount)
 	return header
 }
 
-type GameLump struct {
+type LumpDef struct {
 	Id int32
 	Flags uint16
 	Version uint16
@@ -19,7 +19,7 @@ type GameLump struct {
 	FileLength int32
 }
 
-
+// Note: Nothing below here is actually implemented, as its primarily game/version specific data.
 type StaticPropDictLump struct {
 	DictEntries int32
 	Name []string // Slice length must equal dictEntries. Validation to be added
