+++ lumps/datatypes/model/model.go
@@ -1,0 +1,10 @@
+package model
+
+type Model struct {
+	Mins [3]float32
+	Maxs [3]float32
+	Origin [3]float32
+	HeadNode int32
+	FirstFace int32
+	NumFaces int32
+}
