+++ lumps/datatypes/plane/plane.go
@@ -1,0 +1,9 @@
+package plane
+
+
+type Plane struct {
+	Normal [3]float32 // normal vector
+	Distance float32  // distance from origin
+	AxisType int32	  // plane axis identifier
+}
+
