+++ lumps/datatypes/texdata/texdata.go
@@ -1,0 +1,10 @@
+package texdata
+
+type TexData struct {
+	Reflectivity [3]float32
+	NameStringTableID int32
+	Width int32
+	Height int32
+	ViewWidth int32
+	ViewHeight int32
+}
