+++ lumps/datatypes/texinfo/texinfo.go
@@ -1,0 +1,9 @@
+package texinfo
+
+
+type TexInfo struct {
+	TextureVecs [2][4]float32
+	LightmapVecs [2][4]float32
+	Flags int32
+	TexData int32
+}
