+++ lumps/node.go
@@ -1,7 +1,7 @@
 package lumps
 
 import (
-	"github.com/galaco/bsp/lumps/lumpdata"
+	datatypes "github.com/galaco/bsp/lumps/datatypes/node"
 	"encoding/binary"
 	"bytes"
 	"log"
@@ -13,11 +13,11 @@ import (
  */
 type Node struct {
 	LumpInfo
-	data []lumpdata.Node // MAP_MAX_NODES = 65536
+	data []datatypes.Node // MAP_MAX_NODES = 65536
 }
 
 func (lump Node) FromBytes(raw []byte, length int32) ILump {
-	lump.data = make([]lumpdata.Node, length/int32(unsafe.Sizeof(lumpdata.Node{})))
+	lump.data = make([]datatypes.Node, length/int32(unsafe.Sizeof(datatypes.Node{})))
 	err := binary.Read(bytes.NewBuffer(raw[:]), binary.LittleEndian, &lump.data)
 	if err != nil {
 		log.Fatal(err)
